Separates AppConfig and Application into 'Core' and higher-level classes
This allows us to reuse much of the same config and orchestration code across both iOS apps and app extensions, like a Today View.
