+++ OBAKit/Controls/Tables/TableRowView.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 import AloeStackView
+import OBAKitCore
 
 /// A view that approximates the appearance of a `UITableViewCell`. Meant to be used directly in an `AloeStackView`.
 public class TableRowView: UIView, Highlightable {
