+++ OBAKit/Extensions/AloeStackViewExtensions.swift
@@ -8,6 +8,7 @@
 
 import UIKit
 import AloeStackView
+import OBAKitCore
 
 public extension AloeStackView {
