+++ OBAKitCore/Utilities/Formatters.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 import MapKit
-import OBAKitCore
 
 public class Formatters: NSObject {
     private let locale: Locale
